+++ src/sse/Backend/WebhostingBackend.java
@@ -63,7 +63,7 @@ public class WebhostingBackend implements Backend {
 	}
 
 	@Override
-	public DataOutputStream openBlock(int block) {
+	public DataOutputStream openBlock(long block) {
 
 		if (w != null) {
 			try {
